8371186: Lilliput2: Disable JFR Leak Profiler & related tests until it's fixed with +UseCompactObjectHeaders
The JFR leak profiler stores an encoded reference to a (canonical) path from roots to each of a set of sampled objects after GC at some interval in the space available in the displaced mark word on the sampled object's header. With Lilliput2, there isn't enough space to store such information in the header. Instead, we probably want to use a side hashtable to temporarily keep such information. We've filed a ticket to do that work.

Meanwhile, in this PR, we will disable the Leak Profiler when Lilliput2's `UseCompactObjectHeaders` is enabled. (This is already done in the Leak Profiler, e.g., for Shenanodah GC.)

We'll also disable tests of the LeakProfiler when Lilliput2's UseCompactObjectHeaders is enabled.

The JFR Leak Profiler workaround and the disabled tests will need to be re-enabled once the Leak Profiler is reworked to not use the object header in this manner in Lilliput2 with +UseCompactObjectHeaders.

Reviewed-by: rkennke